SmartClause_2.0/requirements.txt
streamlit>=1.28.0
openai>=1.3.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0
supabase>=2.0.0
python-dotenv>=1.0.0